[NFC] Split `Workspace.swift` into smaller files
