+++ Sources/Workspace/Workspace.swift
@@ -292,35 +292,6 @@ private struct WorkspaceRegistryClientDelegate: RegistryClient.Delegate {
     }
 }
 
-private struct WorkspaceDependencyResolverDelegate: DependencyResolverDelegate {
-    private weak var workspaceDelegate: Workspace.Delegate?
-    private let resolving = ThreadSafeKeyValueStore<PackageIdentity, Bool>()
-
-    init(_ delegate: Workspace.Delegate) {
-        self.workspaceDelegate = delegate
-    }
-
-    func willResolve(term: Term) {
-        // this may be called multiple time by the resolver for various version ranges, but we only want to propagate once since we report at package level
-        resolving.memoize(term.node.package.identity) {
-            self.workspaceDelegate?.willComputeVersion(package: term.node.package.identity, location: term.node.package.locationString)
-            return true
-        }
-    }
-
-    func didResolve(term: Term, version: Version, duration: DispatchTimeInterval) {
-        self.workspaceDelegate?.didComputeVersion(package: term.node.package.identity, location: term.node.package.locationString, version: version.description, duration: duration)
-    }
-
-    // noop
-    func derived(term: Term) {}
-    func conflict(conflict: Incompatibility) {}
-    func satisfied(term: Term, by assignment: Assignment, incompatibility: Incompatibility) {}
-    func partiallySatisfied(term: Term, by assignment: Assignment, incompatibility: Incompatibility, difference: Term) {}
-    func failedToResolve(incompatibility: Incompatibility) {}
-    func solved(result: [(package: PackageReference, binding: BoundVersion, products: ProductFilter)]) {}
-}
-
 private class WorkspaceBinaryArtifactsManagerDelegate: Workspace.BinaryArtifactsManager.Delegate {
     private weak var workspaceDelegate: Workspace.Delegate?
 
@@ -405,30 +376,29 @@ public class Workspace {
     }
 
     /// Source control repository manager used for interacting with source control based dependencies
-    // var for backwards compatibility with deprecated initializers, remove with them
     let repositoryManager: RepositoryManager
 
     /// The registry manager.
-    fileprivate let registryClient: RegistryClient
+    let registryClient: RegistryClient
 
     /// Registry based dependencies downloads manager used for interacting with registry based dependencies
-    fileprivate let registryDownloadsManager: RegistryDownloadsManager
+    let registryDownloadsManager: RegistryDownloadsManager
 
     /// Binary artifacts manager used for downloading and extracting binary artifacts
-    fileprivate let binaryArtifactsManager: BinaryArtifactsManager
+    let binaryArtifactsManager: BinaryArtifactsManager
     
     /// The package fingerprints storage
-    fileprivate let fingerprints: PackageFingerprintStorage?
+    let fingerprints: PackageFingerprintStorage?
 
     /// The workspace configuration settings
-    fileprivate let configuration: WorkspaceConfiguration
+    let configuration: WorkspaceConfiguration
 
-    // state
+    // MARK: State
 
     /// The active package resolver. This is set during a dependency resolution operation.
-    fileprivate var activeResolver: PubGrubDependencyResolver?
+    var activeResolver: PubGrubDependencyResolver?
 
-    fileprivate var resolvedFileWatcher: ResolvedFileWatcher?
+    var resolvedFileWatcher: ResolvedFileWatcher?
 
     /// Create a new package workspace.
     ///
@@ -901,7 +871,12 @@ extension Workspace {
         
         let observabilityScope = observabilityScope.makeChildScope(description: "editing package", metadata: dependency.packageRef.diagnosticsMetadata)
 
-        try self.unedit(dependency: dependency, forceRemove: forceRemove, root: root, observabilityScope: observabilityScope)
+        try self.unedit(
+            dependency: dependency,
+            forceRemove: forceRemove,
+            root: root,
+            observabilityScope: observabilityScope
+        )
     }
 
     /// Perform dependency resolution if needed.
@@ -1097,110 +1072,12 @@ extension Workspace {
         dryRun: Bool = false,
         observabilityScope: ObservabilityScope
     ) throws -> [(PackageReference, Workspace.PackageStateChange)]? {
-        let start = DispatchTime.now()
-        self.delegate?.willUpdateDependencies()
-        defer {
-            self.delegate?.didUpdateDependencies(duration: start.distance(to: .now()))
-        }
-
-        // Create cache directories.
-        self.createCacheDirectories(observabilityScope: observabilityScope)
-
-        // FIXME: this should not block
-        // Load the root manifests and currently checked out manifests.
-        let rootManifests = try temp_await { self.loadRootManifests(packages: root.packages, observabilityScope: observabilityScope, completion: $0) }
-        let rootManifestsMinimumToolsVersion = rootManifests.values.map{ $0.toolsVersion }.min() ?? ToolsVersion.current
-        let resolvedFileOriginHash = try self.computeResolvedFileOriginHash(root: root)
-
-        // Load the current manifests.
-        let graphRoot = PackageGraphRoot(
-            input: root,
-            manifests: rootManifests,
-            dependencyMapper: self.dependencyMapper,
+        try self._updateDependencies(
+            root: root,
+            packages: packages,
+            dryRun: dryRun,
             observabilityScope: observabilityScope
         )
-        let currentManifests = try self.loadDependencyManifests(root: graphRoot, observabilityScope: observabilityScope)
-
-        // Abort if we're unable to load the pinsStore or have any diagnostics.
-        guard let pinsStore = observabilityScope.trap({ try self.pinsStore.load() }) else { return nil }
-
-        // Ensure we don't have any error at this point.
-        guard !observabilityScope.errorsReported else {
-            return nil
-        }
-
-        // Add unversioned constraints for edited packages.
-        var updateConstraints = currentManifests.editedPackagesConstraints
-
-        // Create constraints based on root manifest and pins for the update resolution.
-        updateConstraints += try graphRoot.constraints()
-
-        let pins: PinsStore.Pins
-        if packages.isEmpty {
-            // No input packages so we have to do a full update. Set pins map to empty.
-            pins = [:]
-        } else {
-            // We have input packages so we have to partially update the package graph. Remove
-            // the pins for the input packages so only those packages are updated.
-            pins = pinsStore.pins.filter{ !packages.contains($0.value.packageRef.identity.description) && !packages.contains($0.value.packageRef.deprecatedName) }
-        }
-
-        // Resolve the dependencies.
-        let resolver = try self.createResolver(pins: pins, observabilityScope: observabilityScope)
-        self.activeResolver = resolver
-
-        let updateResults = self.resolveDependencies(
-            resolver: resolver,
-            constraints: updateConstraints,
-            observabilityScope: observabilityScope
-        )
-
-        // Reset the active resolver.
-        self.activeResolver = nil
-
-        guard !observabilityScope.errorsReported else {
-            return nil
-        }
-
-        if dryRun {
-            return observabilityScope.trap {
-                return try self.computePackageStateChanges(root: graphRoot, resolvedDependencies: updateResults, updateBranches: true, observabilityScope: observabilityScope)
-            }
-        }
-
-        // Update the checkouts based on new dependency resolution.
-        let packageStateChanges = self.updateDependenciesCheckouts(root: graphRoot, updateResults: updateResults, updateBranches: true, observabilityScope: observabilityScope)
-        guard !observabilityScope.errorsReported else {
-            return nil
-        }
-
-        // Load the updated manifests.
-        let updatedDependencyManifests = try self.loadDependencyManifests(root: graphRoot, observabilityScope: observabilityScope)
-        // If we have missing packages, something is fundamentally wrong with the resolution of the graph
-        let stillMissingPackages = try updatedDependencyManifests.missingPackages
-        guard stillMissingPackages.isEmpty else {
-            observabilityScope.emit(.exhaustedAttempts(missing: stillMissingPackages))
-            return nil
-        }
-
-        // Update the resolved file.
-        try self.saveResolvedFile(
-            pinsStore: pinsStore,
-            dependencyManifests: updatedDependencyManifests,
-            originHash: resolvedFileOriginHash,
-            rootManifestsMinimumToolsVersion: rootManifestsMinimumToolsVersion,
-            observabilityScope: observabilityScope
-        )
-
-        // Update the binary target artifacts.
-        let addedOrUpdatedPackages = packageStateChanges.compactMap({ $0.1.isAddedOrUpdated ? $0.0 : nil })
-        try self.updateBinaryArtifacts(
-            manifests: updatedDependencyManifests,
-            addedOrUpdatedPackages: addedOrUpdatedPackages,
-            observabilityScope: observabilityScope
-        )
-
-        return packageStateChanges
     }
 
     @discardableResult
@@ -1473,1431 +1350,29 @@ extension Workspace {
             fileRuleDescription.match(path: filePath, toolsVersion: self.currentToolsVersion)
         }
     }
-}
 
-
-// MARK: - Editing Functions
-
-extension Workspace {
-    /// Edit implementation.
-    fileprivate func _edit(
-        packageName: String,
-        path: AbsolutePath? = nil,
-        revision: Revision? = nil,
-        checkoutBranch: String? = nil,
-        observabilityScope: ObservabilityScope
-    ) throws {
-        // Look up the dependency and check if we can edit it.
-        guard let dependency = self.state.dependencies[.plain(packageName)] else {
-            observabilityScope.emit(.dependencyNotFound(packageName: packageName))
-            return
-        }
-        
-        let observabilityScope = observabilityScope.makeChildScope(description: "editing package", metadata: dependency.packageRef.diagnosticsMetadata)
-
-        let checkoutState: CheckoutState
-        switch dependency.state {
-        case .sourceControlCheckout(let _checkoutState):
-            checkoutState = _checkoutState
-        case .edited:
-            observabilityScope.emit(error: "dependency '\(dependency.packageRef.identity)' already in edit mode")
-            return
-        case .fileSystem:
-            observabilityScope.emit(error: "local dependency '\(dependency.packageRef.identity)' can't be edited")
-            return
-        case .registryDownload:
-            observabilityScope.emit(error: "registry dependency '\(dependency.packageRef.identity)' can't be edited")
-            return
-        case .custom:
-            observabilityScope.emit(error: "custom dependency '\(dependency.packageRef.identity)' can't be edited")
-            return
-        }
-
-        // If a path is provided then we use it as destination. If not, we
-        // use the folder with packageName inside editablesPath.
-        let destination = path ?? self.location.editsDirectory.appending(component: packageName)
-
-        // If there is something present at the destination, we confirm it has
-        // a valid manifest with name same as the package we are trying to edit.
-        if fileSystem.exists(destination) {
-            // FIXME: this should not block
-            let manifest = try temp_await {
-                self.loadManifest(packageIdentity: dependency.packageRef.identity,
-                                  packageKind: .fileSystem(destination),
-                                  packagePath: destination,
-                                  packageLocation: dependency.packageRef.locationString,
-                                  observabilityScope: observabilityScope,
-                                  completion: $0)
-            }
-
-            guard manifest.displayName == packageName else {
-                return observabilityScope.emit(error: "package at '\(destination)' is \(manifest.displayName) but was expecting \(packageName)")
-            }
-
-            // Emit warnings for branch and revision, if they're present.
-            if let checkoutBranch {
-                observabilityScope.emit(.editBranchNotCheckedOut(
-                    packageName: packageName,
-                    branchName: checkoutBranch))
-            }
-            if let revision {
-                observabilityScope.emit(.editRevisionNotUsed(
-                    packageName: packageName,
-                    revisionIdentifier: revision.identifier))
-            }
-        } else {
-            // Otherwise, create a checkout at the destination from our repository store.
-            //
-            // Get handle to the repository.
-            // TODO: replace with async/await when available
-            let repository = try dependency.packageRef.makeRepositorySpecifier()
-            let handle = try temp_await {
-                repositoryManager.lookup(
-                    package: dependency.packageRef.identity,
-                    repository: repository,
-                    updateStrategy: .never,
-                    observabilityScope: observabilityScope,
-                    delegateQueue: .sharedConcurrent,
-                    callbackQueue: .sharedConcurrent,
-                    completion: $0
-                )
-            }
-            let repo = try handle.open()
-
-            // Do preliminary checks on branch and revision, if provided.
-            if let branch = checkoutBranch, repo.exists(revision: Revision(identifier: branch)) {
-                throw WorkspaceDiagnostics.BranchAlreadyExists(branch: branch)
-            }
-            if let revision, !repo.exists(revision: revision) {
-                throw WorkspaceDiagnostics.RevisionDoesNotExist(revision: revision.identifier)
-            }
-
-            let workingCopy = try handle.createWorkingCopy(at: destination, editable: true)
-            try workingCopy.checkout(revision: revision ?? checkoutState.revision)
-
-            // Checkout to the new branch if provided.
-            if let branch = checkoutBranch {
-                try workingCopy.checkout(newBranch: branch)
-            }
-        }
-
-        // For unmanaged dependencies, create the symlink under editables dir.
-        if let path {
-            try fileSystem.createDirectory(self.location.editsDirectory)
-            // FIXME: We need this to work with InMem file system too.
-            if !(fileSystem is InMemoryFileSystem) {
-                let symLinkPath = self.location.editsDirectory.appending(component: packageName)
-
-                // Cleanup any existing symlink.
-                if fileSystem.isSymlink(symLinkPath) {
-                    try fileSystem.removeFileTree(symLinkPath)
-                }
-
-                // FIXME: We should probably just warn in case we fail to create
-                // this symlink, which could happen if there is some non-symlink
-                // entry at this location.
-                try fileSystem.createSymbolicLink(symLinkPath, pointingAt: path, relative: false)
-            }
-        }
-
-        // Remove the existing checkout.
-        do {
-            let oldCheckoutPath = self.location.repositoriesCheckoutSubdirectory(for: dependency)
-            try fileSystem.chmod(.userWritable, path: oldCheckoutPath, options: [.recursive, .onlyFiles])
-            try fileSystem.removeFileTree(oldCheckoutPath)
-        }
-
-        // Save the new state.
-        self.state.dependencies.add(
-            try dependency.edited(subpath: RelativePath(validating: packageName), unmanagedPath: path)
-        )
-        try self.state.save()
-    }
-
-    /// Unedit a managed dependency. See public API unedit(packageName:forceRemove:).
-    func unedit(
-        dependency: ManagedDependency,
-        forceRemove: Bool,
-        root: PackageGraphRootInput? = nil,
-        observabilityScope: ObservabilityScope
-    ) throws {
-
-        // Compute if we need to force remove.
-        var forceRemove = forceRemove
-
-        // If the dependency isn't in edit mode, we can't unedit it.
-        guard case .edited(_, let unmanagedPath) = dependency.state else {
-            throw WorkspaceDiagnostics.DependencyNotInEditMode(dependencyName: dependency.packageRef.identity.description)
-        }
-
-        // Set force remove to true for unmanaged dependencies.  Note that
-        // this only removes the symlink under the editable directory and
-        // not the actual unmanaged package.
-        if unmanagedPath != nil {
-            forceRemove = true
-        }
-
-        // Form the edit working repo path.
-        let path = self.location.editSubdirectory(for: dependency)
-        // Check for uncommited and unpushed changes if force removal is off.
-        if !forceRemove {
-            let workingCopy = try repositoryManager.openWorkingCopy(at: path)
-            guard !workingCopy.hasUncommittedChanges() else {
-                throw WorkspaceDiagnostics.UncommitedChanges(repositoryPath: path)
-            }
-            guard try !workingCopy.hasUnpushedCommits() else {
-                throw WorkspaceDiagnostics.UnpushedChanges(repositoryPath: path)
-            }
-        }
-        // Remove the editable checkout from disk.
-        if fileSystem.exists(path) {
-            try fileSystem.removeFileTree(path)
-        }
-        // If this was the last editable dependency, remove the editables directory too.
-        if fileSystem.exists(self.location.editsDirectory), try fileSystem.getDirectoryContents(self.location.editsDirectory).isEmpty {
-            try fileSystem.removeFileTree(self.location.editsDirectory)
-        }
-
-        if case .edited(let basedOn, _) = dependency.state, case .sourceControlCheckout(let checkoutState) = basedOn?.state {
-            // Restore the original checkout.
-            //
-            // The retrieve method will automatically update the managed dependency state.
-            _ = try self.checkoutRepository(package: dependency.packageRef, at: checkoutState, observabilityScope: observabilityScope)
-        } else {
-            // The original dependency was removed, update the managed dependency state.
-            self.state.dependencies.remove(dependency.packageRef.identity)
-            try self.state.save()
-        }
-
-        // Resolve the dependencies if workspace root is provided. We do this to
-        // ensure the unedited version of this dependency is resolved properly.
-        if let root {
-            try self._resolve(
-                root: root,
-                explicitProduct: .none,
-                resolvedFileStrategy: .update(forceResolution: false),
-                observabilityScope: observabilityScope
-            )
-        }
-    }
-
-    /// Create the cache directories.
-    fileprivate func createCacheDirectories(observabilityScope: ObservabilityScope) {
-        observabilityScope.trap {
-            try fileSystem.createDirectory(self.repositoryManager.path, recursive: true)
-            try fileSystem.createDirectory(self.location.repositoriesCheckoutsDirectory, recursive: true)
-            try fileSystem.createDirectory(self.location.artifactsDirectory, recursive: true)
-        }
-    }
-}
-
-// MARK: - Pinning Functions
-
-extension Workspace {
-    /// Pins all of the current managed dependencies at their checkout state.
-    fileprivate func saveResolvedFile(
-        pinsStore: PinsStore,
-        dependencyManifests: DependencyManifests,
-        originHash: String,
-        rootManifestsMinimumToolsVersion: ToolsVersion,
-        observabilityScope: ObservabilityScope
-    ) throws {
-        var dependenciesToPin = [ManagedDependency]()
-        let requiredDependencies = try dependencyManifests.requiredPackages.filter({ $0.kind.isPinnable })
-        for dependency in requiredDependencies {
-            if let managedDependency = self.state.dependencies[comparingLocation: dependency] {
-                dependenciesToPin.append(managedDependency)
-            } else {
-                observabilityScope.emit(warning: "required dependency \(dependency.identity) (\(dependency.locationString)) was not found in managed dependencies and will not be recorded in resolved file")
-            }
-        }
-
-        // try to load the pin store from disk so we can compare for any changes
-        // this is needed as we want to avoid re-writing the resolved files unless absolutely necessary
-        var needsUpdate = false
-        if let storedPinStore = try? self.pinsStore.load() {
-            // compare for any differences between the existing state and the stored one
-            // subtle changes between versions of SwiftPM could treat URLs differently
-            // in which case we don't want to cause unnecessary churn
-            if dependenciesToPin.count != storedPinStore.pins.count {
-                needsUpdate = true
-            } else {
-                for dependency in dependenciesToPin {
-                    if let pin = storedPinStore.pins.first(where: { $0.value.packageRef.equalsIncludingLocation(dependency.packageRef) }) {
-                        if pin.value.state != PinsStore.Pin(dependency)?.state {
-                            needsUpdate = true
-                            break
-                        }
-                    } else {
-                        needsUpdate = true
-                        break
-                    }
-                }
-            }
-        } else {
-            needsUpdate = true
-        }
-
-        // exist early is there is nothing to do
-        if !needsUpdate {
-            return
-        }
-
-        // reset the pinsStore and start pinning the required dependencies.
-        pinsStore.unpinAll()
-        for dependency in dependenciesToPin {
-            pinsStore.pin(dependency)
-        }
-
-        observabilityScope.trap {
-            try pinsStore.saveState(
-                toolsVersion: rootManifestsMinimumToolsVersion,
-                originHash: originHash
-            )
-        }
-
-        // Ask resolved file watcher to update its value so we don't fire
-        // an extra event if the file was modified by us.
-        self.resolvedFileWatcher?.updateValue()
-    }
-
-    /// Watch the Package.resolved for changes.
-    ///
-    /// This is useful if clients want to be notified when the Package.resolved
-    /// file is changed *outside* of libSwiftPM operations. For example, as part
-    /// of a git operation.
-    public func watchResolvedFile() throws {
-        // Return if we're already watching it.
-        guard self.resolvedFileWatcher == nil else { return }
-        self.resolvedFileWatcher = try ResolvedFileWatcher(resolvedFile: self.location.resolvedVersionsFile) { [weak self] in
-            self?.delegate?.resolvedFileChanged()
-        }
-    }
-}
-
-fileprivate extension PinsStore {
-    /// Pin a managed dependency at its checkout state.
-    ///
-    /// This method does nothing if the dependency is in edited state.
-    func pin(_ dependency: Workspace.ManagedDependency) {
-        if let pin = PinsStore.Pin(dependency) {
-            self.add(pin)
-        }
-    }
-}
-
-fileprivate extension PinsStore.Pin {
-    init?(_ dependency: Workspace.ManagedDependency) {
-        switch dependency.state {
-        case .sourceControlCheckout(.version(let version, let revision)):
-            self.init(
-                packageRef: dependency.packageRef,
-                state: .version(version, revision: revision.identifier)
-            )
-        case .sourceControlCheckout(.branch(let branch, let revision)):
-            self.init(
-                packageRef: dependency.packageRef,
-                state: .branch(name: branch, revision: revision.identifier)
-            )
-        case .sourceControlCheckout(.revision(let revision)):
-            self.init(
-                packageRef: dependency.packageRef,
-                state: .revision(revision.identifier)
-            )
-        case .registryDownload(let version):
-            self.init(
-                packageRef: dependency.packageRef,
-                state: .version(version, revision: .none)
-            )
-        case .edited, .fileSystem, .custom:
-            // NOOP
-            return nil
-        }
-    }
-}
-
-
-// MARK: - Binary artifacts
-
-extension Workspace {
-    fileprivate func updateBinaryArtifacts(
-        manifests: DependencyManifests,
-        addedOrUpdatedPackages: [PackageReference],
-        observabilityScope: ObservabilityScope
-    ) throws {
-        let manifestArtifacts = try self.binaryArtifactsManager.parseArtifacts(from: manifests, observabilityScope: observabilityScope)
-
-        var artifactsToRemove: [ManagedArtifact] = []
-        var artifactsToAdd: [ManagedArtifact] = []
-        var artifactsToDownload: [BinaryArtifactsManager.RemoteArtifact] = []
-        var artifactsToExtract: [ManagedArtifact] = []
-
-        for artifact in state.artifacts {
-            if !manifestArtifacts.local.contains(where: { $0.packageRef == artifact.packageRef && $0.targetName == artifact.targetName }) &&
-                !manifestArtifacts.remote.contains(where: { $0.packageRef == artifact.packageRef && $0.targetName == artifact.targetName }) {
-                artifactsToRemove.append(artifact)
-            }
-        }
-
-        for artifact in manifestArtifacts.local {
-            let existingArtifact = self.state.artifacts[
-                packageIdentity: artifact.packageRef.identity,
-                targetName: artifact.targetName
-            ]
-
-            if artifact.path.extension?.lowercased() == "zip" {
-                // If we already have an artifact that was extracted from an archive with the same checksum,
-                // we don't need to extract the artifact again.
-                if case .local(let existingChecksum) = existingArtifact?.source, existingChecksum == (try self.binaryArtifactsManager.checksum(forBinaryArtifactAt: artifact.path)) {
-                    continue
-                }
-
-                artifactsToExtract.append(artifact)
-            } else {
-                guard let _ = try BinaryArtifactsManager.deriveBinaryArtifact(fileSystem: self.fileSystem, path: artifact.path, observabilityScope: observabilityScope) else {
-                    observabilityScope.emit(.localArtifactNotFound(artifactPath: artifact.path, targetName: artifact.targetName))
-                    continue
-                }
-                artifactsToAdd.append(artifact)
-            }
-
-            if let existingArtifact, isAtArtifactsDirectory(existingArtifact) {
-                // Remove the old extracted artifact, be it local archived or remote one.
-                artifactsToRemove.append(existingArtifact)
-            }
-        }
-
-        for artifact in manifestArtifacts.remote {
-            let existingArtifact = self.state.artifacts[
-                packageIdentity: artifact.packageRef.identity,
-                targetName: artifact.targetName
-            ]
-
-            if let existingArtifact {
-                if case .remote(let existingURL, let existingChecksum) = existingArtifact.source {
-                    // If we already have an artifact with the same checksum, we don't need to download it again.
-                    if artifact.checksum == existingChecksum {
-                        continue
-                    }
-
-                    let urlChanged = artifact.url != URL(string: existingURL)
-                    // If the checksum is different but the package wasn't updated, this is a security risk.
-                    if !urlChanged && !addedOrUpdatedPackages.contains(artifact.packageRef) {
-                        observabilityScope.emit(.artifactChecksumChanged(targetName: artifact.targetName))
-                        continue
-                    }
-                }
-
-                if isAtArtifactsDirectory(existingArtifact) {
-                    // Remove the old extracted artifact, be it local archived or remote one.
-                    artifactsToRemove.append(existingArtifact)
-                }
-            }
-
-            artifactsToDownload.append(artifact)
-        }
-
-        // Remove the artifacts and directories which are not needed anymore.
-        observabilityScope.trap {
-            for artifact in artifactsToRemove {
-                state.artifacts.remove(packageIdentity: artifact.packageRef.identity, targetName: artifact.targetName)
-
-                if isAtArtifactsDirectory(artifact) {
-                    try fileSystem.removeFileTree(artifact.path)
-                }
-            }
-
-            for directory in try fileSystem.getDirectoryContents(self.location.artifactsDirectory) {
-                let directoryPath = self.location.artifactsDirectory.appending(component: directory)
-                if try fileSystem.isDirectory(directoryPath) && fileSystem.getDirectoryContents(directoryPath).isEmpty {
-                    try fileSystem.removeFileTree(directoryPath)
-                }
-            }
-        }
-
-        guard !observabilityScope.errorsReported else {
-            throw Diagnostics.fatalError
-        }
-
-        // Download the artifacts
-        let downloadedArtifacts = try self.binaryArtifactsManager.download(
-            artifactsToDownload,
-            artifactsDirectory: self.location.artifactsDirectory,
-            observabilityScope: observabilityScope
-        )
-        artifactsToAdd.append(contentsOf: downloadedArtifacts)
-
-        // Extract the local archived artifacts
-        let extractedLocalArtifacts = try self.binaryArtifactsManager.extract(
-            artifactsToExtract,
-            artifactsDirectory: self.location.artifactsDirectory,
-            observabilityScope: observabilityScope
-        )
-        artifactsToAdd.append(contentsOf: extractedLocalArtifacts)
-
-        // Add the new artifacts
-        for artifact in artifactsToAdd {
-            self.state.artifacts.add(artifact)
-        }
-
-        guard !observabilityScope.errorsReported else {
-            throw Diagnostics.fatalError
-        }
-
-        observabilityScope.trap {
-            try self.state.save()
-        }
-
-        func isAtArtifactsDirectory(_ artifact: ManagedArtifact) -> Bool {
-            artifact.path.isDescendant(of: self.location.artifactsDirectory)
-        }
-    }
-}
-
-// MARK: - Dependency Management
-
-extension Workspace {
-    enum ResolvedFileStrategy {
-        case lockFile
-        case update(forceResolution: Bool)
-        case bestEffort
-    }
-
-    @discardableResult
-    private func _resolve(
-        root: PackageGraphRootInput,
-        explicitProduct: String?,
-        resolvedFileStrategy: ResolvedFileStrategy,
-        observabilityScope: ObservabilityScope
-    ) throws -> DependencyManifests {
-        let start = DispatchTime.now()
-        self.delegate?.willResolveDependencies()
-        defer {
-            self.delegate?.didResolveDependencies(duration: start.distance(to: .now()))
-        }
-
-        switch resolvedFileStrategy {
-        case .lockFile:
-            observabilityScope.emit(info: "using '\(self.location.resolvedVersionsFile.basename)' file as lock file")
-            return try self._resolveBasedOnResolvedVersionsFile(
-                root: root,
-                explicitProduct: explicitProduct,
-                observabilityScope: observabilityScope
-            )
-        case .update(let forceResolution):
-            return try resolveAndUpdateResolvedFile(forceResolution: forceResolution)
-        case .bestEffort:
-            guard !self.state.dependencies.hasEditedDependencies() else {
-                return try resolveAndUpdateResolvedFile(forceResolution: false)
-            }
-            guard self.fileSystem.exists(self.location.resolvedVersionsFile) else {
-                return try resolveAndUpdateResolvedFile(forceResolution: false)
-            }
-
-            guard let pinsStore = try? self.pinsStore.load(), let storedHash = pinsStore.originHash else {
-                observabilityScope.emit(debug: "'\(self.location.resolvedVersionsFile.basename)' origin hash is missing. resolving and updating accordingly")
-                return try resolveAndUpdateResolvedFile(forceResolution: false)
-            }
-
-            let currentHash = try self.computeResolvedFileOriginHash(root: root)
-            guard storedHash == currentHash else {
-                observabilityScope.emit(debug: "'\(self.location.resolvedVersionsFile.basename)' origin hash does do not match manifest dependencies. resolving and updating accordingly")
-                return try resolveAndUpdateResolvedFile(forceResolution: false)
-            }
-
-            observabilityScope.emit(debug: "'\(self.location.resolvedVersionsFile.basename)' origin hash matches manifest dependencies, attempting resolution based on this file")
-            let (manifests, precomputationResult) = try self.tryResolveBasedOnResolvedVersionsFile(
-                root: root,
-                explicitProduct: explicitProduct,
-                observabilityScope: observabilityScope
-            )
-            switch precomputationResult {
-            case .notRequired:
-                return manifests
-            case .required(reason: .errorsPreviouslyReported):
-                return manifests
-            case .required(let reason):
-                // FIXME: ideally this is not done based on a side-effect
-                let reasonString = Self.format(workspaceResolveReason: reason)
-                observabilityScope.emit(debug: "resolution based on '\(self.location.resolvedVersionsFile.basename)' could not be completed because \(reasonString). resolving and updating accordingly")
-                return try resolveAndUpdateResolvedFile(forceResolution: false)
-            }
-        }
-
-        func resolveAndUpdateResolvedFile(forceResolution: Bool) throws -> DependencyManifests {
-            observabilityScope.emit(debug: "resolving and updating '\(self.location.resolvedVersionsFile.basename)'")
-            return try self.resolveAndUpdateResolvedFile(
-                root: root,
-                explicitProduct: explicitProduct,
-                forceResolution: forceResolution,
-                constraints: [],
-                observabilityScope: observabilityScope
-            )
-        }
-    }
-
-    private func computeResolvedFileOriginHash(root: PackageGraphRootInput) throws -> String {
-        var content = try root.packages.reduce(into: "", { partial, element in
-            let path = try ManifestLoader.findManifest(
-                packagePath: element,
-                fileSystem: self.fileSystem,
-                currentToolsVersion: self.currentToolsVersion
-            )
-            partial.append(try self.fileSystem.readFileContents(path))
-        })
-        content += root.dependencies.reduce(into: "", { partial, element in
-            partial += element.locationString
-        })
-        return content.sha256Checksum
-    }
-
-    @discardableResult
-    func _resolveBasedOnResolvedVersionsFile(
-        root: PackageGraphRootInput,
-        explicitProduct: String?,
-        observabilityScope: ObservabilityScope) throws -> DependencyManifests {
-        let (manifests, precomputationResult) = try self.tryResolveBasedOnResolvedVersionsFile(
-            root: root,
-            explicitProduct: explicitProduct,
-            observabilityScope: observabilityScope
-        )
-        switch precomputationResult {
-        case .notRequired:
-            return manifests
-        case .required(reason: .errorsPreviouslyReported):
-            return manifests
-        case .required(let reason):
-            // FIXME: ideally this is not done based on a side-effect
-            let reasonString = Self.format(workspaceResolveReason: reason)
-            if !self.fileSystem.exists(self.location.resolvedVersionsFile) {
-                observabilityScope.emit(error: "a resolved file is required when automatic dependency resolution is disabled and should be placed at \(self.location.resolvedVersionsFile.pathString). \(reasonString)")
-            } else {
-                observabilityScope.emit(error: "an out-of-date resolved file was detected at \(self.location.resolvedVersionsFile.pathString), which is not allowed when automatic dependency resolution is disabled; please make sure to update the file to reflect the changes in dependencies. \(reasonString)")
-            }
-            return manifests
-        }
-    }
-
-    /// Resolves the dependencies according to the entries present in the Package.resolved file.
-    ///
-    /// This method bypasses the dependency resolution and resolves dependencies
-    /// according to the information in the resolved file.
-    fileprivate func tryResolveBasedOnResolvedVersionsFile(
-        root: PackageGraphRootInput,
-        explicitProduct: String?,
-        observabilityScope: ObservabilityScope
-    ) throws -> (DependencyManifests, ResolutionPrecomputationResult) {
-        // Ensure the cache path exists.
-        self.createCacheDirectories(observabilityScope: observabilityScope)
-
-        // FIXME: this should not block
-        let rootManifests = try temp_await { self.loadRootManifests(packages: root.packages, observabilityScope: observabilityScope, completion: $0) }
-        let graphRoot = PackageGraphRoot(
-            input: root,
-            manifests: rootManifests,
-            explicitProduct: explicitProduct,
-            dependencyMapper: self.dependencyMapper,
-            observabilityScope: observabilityScope
-        )
-
-        // Load the pins store or abort now.
-        guard let pinsStore = observabilityScope.trap({ try self.pinsStore.load() }), !observabilityScope.errorsReported else {
-            return (try self.loadDependencyManifests(root: graphRoot, observabilityScope: observabilityScope), .notRequired)
-        }
-
-        // Request all the containers to fetch them in parallel.
-        //
-        // We just request the packages here, repository manager will
-        // automatically manage the parallelism.
-        let group = DispatchGroup()
-        for pin in pinsStore.pins.values {
-            group.enter()
-            let observabilityScope = observabilityScope.makeChildScope(description: "requesting package containers", metadata: pin.packageRef.diagnosticsMetadata)
-
-            let updateStrategy: ContainerUpdateStrategy = {
-                if self.configuration.skipDependenciesUpdates {
-                    return .never
-                } else {
-                    switch pin.state {
-                    case .branch(_, let revision):
-                        return .ifNeeded(revision: revision)
-                    case .revision(let revision):
-                        return .ifNeeded(revision: revision)
-                    case .version(_, let .some(revision)):
-                        return .ifNeeded(revision: revision)
-                    case .version(_, .none):
-                        return .always
-                    }
-                }
-            }()
-
-            self.packageContainerProvider.getContainer(
-                for: pin.packageRef,
-                updateStrategy:  updateStrategy,
-                observabilityScope: observabilityScope,
-                on: .sharedConcurrent,
-                completion: { _ in group.leave() }
-            )
-        }
-        group.wait()
-
-        // Compute the pins that we need to actually clone.
-        //
-        // We require cloning if there is no checkout or if the checkout doesn't
-        // match with the pin.
-        let requiredPins = pinsStore.pins.values.filter{ pin in
-            // also compare the location in case it has changed
-            guard let dependency = state.dependencies[comparingLocation: pin.packageRef] else {
-                return true
-            }
-            switch dependency.state {
-            case .sourceControlCheckout(let checkoutState):
-                return !pin.state.equals(checkoutState)
-            case .registryDownload(let version):
-                return !pin.state.equals(version)
-            case .edited, .fileSystem, .custom:
-                return true
-            }
-        }
-
-        // Retrieve the required pins.
-        for pin in requiredPins {
-            observabilityScope.makeChildScope(description: "retrieving dependency pins", metadata: pin.packageRef.diagnosticsMetadata).trap {
-                switch pin.packageRef.kind {
-                case .localSourceControl, .remoteSourceControl:
-                    _ = try self.checkoutRepository(package: pin.packageRef, at: pin.state, observabilityScope: observabilityScope)
-                case .registry:
-                    _ = try self.downloadRegistryArchive(package: pin.packageRef, at: pin.state, observabilityScope: observabilityScope)
-                default:
-                    throw InternalError("invalid pin type \(pin.packageRef.kind)")
-                }
-            }
-        }
-
-        let currentManifests = try self.loadDependencyManifests(root: graphRoot, automaticallyAddManagedDependencies: true, observabilityScope: observabilityScope)
-
-        try self.updateBinaryArtifacts(manifests: currentManifests, addedOrUpdatedPackages: [], observabilityScope: observabilityScope)
-
-        let precomputationResult = try self.precomputeResolution(
-            root: graphRoot,
-            dependencyManifests: currentManifests,
-            pinsStore: pinsStore,
-            constraints: [],
-            observabilityScope: observabilityScope
-        )
-
-        return (currentManifests, precomputationResult)
-    }
-
-    /// Implementation of resolve(root:diagnostics:).
-    ///
-    /// The extra constraints will be added to the main requirements.
-    /// It is useful in situations where a requirement is being
-    /// imposed outside of manifest and pins file. E.g., when using a command
-    /// like `$ swift package resolve foo --version 1.0.0`.
-    @discardableResult
-    fileprivate func resolveAndUpdateResolvedFile(
-        root: PackageGraphRootInput,
-        explicitProduct: String? = nil,
-        forceResolution: Bool,
-        constraints: [PackageContainerConstraint],
-        observabilityScope: ObservabilityScope
-    ) throws -> DependencyManifests {
-        // Ensure the cache path exists and validate that edited dependencies.
-        self.createCacheDirectories(observabilityScope: observabilityScope)
-
-        // FIXME: this should not block
-        // Load the root manifests and currently checked out manifests.
-        let rootManifests = try temp_await { self.loadRootManifests(packages: root.packages, observabilityScope: observabilityScope, completion: $0) }
-        let rootManifestsMinimumToolsVersion = rootManifests.values.map{ $0.toolsVersion }.min() ?? ToolsVersion.current
-        let resolvedFileOriginHash = try self.computeResolvedFileOriginHash(root: root)
-
-        // Load the current manifests.
-        let graphRoot = PackageGraphRoot(
-            input: root,
-            manifests: rootManifests,
-            explicitProduct: explicitProduct,
-            dependencyMapper: self.dependencyMapper,
-            observabilityScope: observabilityScope
-        )
-        let currentManifests = try self.loadDependencyManifests(root: graphRoot, observabilityScope: observabilityScope)
-        guard !observabilityScope.errorsReported else {
-            return currentManifests
-        }
-
-        // load and update the pins store with any changes from loading the top level dependencies
-        guard let pinsStore = self.loadAndUpdatePinsStore(
-            dependencyManifests: currentManifests,
-            rootManifestsMinimumToolsVersion: rootManifestsMinimumToolsVersion,
-            observabilityScope: observabilityScope
-        ) else {
-            // abort if PinsStore reported any errors.
-            return currentManifests
-        }
-
-        // abort if PinsStore reported any errors.
-        guard !observabilityScope.errorsReported else {
-            return currentManifests
-        }
-
-        // Compute the missing package identities.
-        let missingPackages = try currentManifests.missingPackages
-
-        // Compute if we need to run the resolver. We always run the resolver if
-        // there are extra constraints.
-        if !missingPackages.isEmpty {
-            delegate?.willResolveDependencies(reason: .newPackages(packages: Array(missingPackages)))
-        } else if !constraints.isEmpty || forceResolution {
-            delegate?.willResolveDependencies(reason: .forced)
-        } else {
-            let result = try self.precomputeResolution(
-                root: graphRoot,
-                dependencyManifests: currentManifests,
-                pinsStore: pinsStore,
-                constraints: constraints,
-                observabilityScope: observabilityScope
-            )
-
-            switch result {
-            case .notRequired:
-                // since nothing changed we can exit early,
-                // but need update resolved file and download an missing binary artifact
-                try self.saveResolvedFile(
-                    pinsStore: pinsStore,
-                    dependencyManifests: currentManifests,
-                    originHash: resolvedFileOriginHash,
-                    rootManifestsMinimumToolsVersion: rootManifestsMinimumToolsVersion,
-                    observabilityScope: observabilityScope
-                )
-
-                try self.updateBinaryArtifacts(
-                    manifests: currentManifests,
-                    addedOrUpdatedPackages: [],
-                    observabilityScope: observabilityScope
-                )
-
-                return currentManifests
-            case .required(let reason):
-                delegate?.willResolveDependencies(reason: reason)
-            }
-        }
-
-        // Create the constraints.
-        var computedConstraints = [PackageContainerConstraint]()
-        computedConstraints += currentManifests.editedPackagesConstraints
-        computedConstraints += try graphRoot.constraints() + constraints
-
-        // Perform dependency resolution.
-        let resolver = try self.createResolver(pins: pinsStore.pins, observabilityScope: observabilityScope)
-        self.activeResolver = resolver
-
-        let result = self.resolveDependencies(
-            resolver: resolver,
-            constraints: computedConstraints,
-            observabilityScope: observabilityScope
-        )
-
-        // Reset the active resolver.
-        self.activeResolver = nil
-
-        guard !observabilityScope.errorsReported else {
-            return currentManifests
-        }
-
-        // Update the checkouts with dependency resolution result.
-        let packageStateChanges = self.updateDependenciesCheckouts(root: graphRoot, updateResults: result, observabilityScope: observabilityScope)
-        guard !observabilityScope.errorsReported else {
-            return currentManifests
-        }
-
-        // Update the pinsStore.
-        let updatedDependencyManifests = try self.loadDependencyManifests(root: graphRoot, observabilityScope: observabilityScope)
-        // If we still have missing packages, something is fundamentally wrong with the resolution of the graph
-        let stillMissingPackages = try updatedDependencyManifests.missingPackages
-        guard stillMissingPackages.isEmpty else {
-            observabilityScope.emit(.exhaustedAttempts(missing: stillMissingPackages))
-            return updatedDependencyManifests
-        }
-
-        // Update the resolved file.
-        try self.saveResolvedFile(
-            pinsStore: pinsStore,
-            dependencyManifests: updatedDependencyManifests,
-            originHash: resolvedFileOriginHash,
-            rootManifestsMinimumToolsVersion: rootManifestsMinimumToolsVersion,
-            observabilityScope: observabilityScope
-        )
-
-        let addedOrUpdatedPackages = packageStateChanges.compactMap({ $0.1.isAddedOrUpdated ? $0.0 : nil })
-        try self.updateBinaryArtifacts(
-            manifests: updatedDependencyManifests,
-            addedOrUpdatedPackages: addedOrUpdatedPackages,
-            observabilityScope: observabilityScope
-        )
-
-        return updatedDependencyManifests
-    }
-
-
-    /// Updates the current working checkouts i.e. clone or remove based on the
-    /// provided dependency resolution result.
-    ///
-    /// - Parameters:
-    ///   - updateResults: The updated results from dependency resolution.
-    ///   - diagnostics: The diagnostics engine that reports errors, warnings
-    ///     and notes.
-    ///   - updateBranches: If the branches should be updated in case they're pinned.
-    @discardableResult
-    fileprivate func updateDependenciesCheckouts(
-        root: PackageGraphRoot,
-        updateResults: [(PackageReference, BoundVersion, ProductFilter)],
-        updateBranches: Bool = false,
-        observabilityScope: ObservabilityScope
-    ) -> [(PackageReference, PackageStateChange)] {
-        // Get the update package states from resolved results.
-        guard let packageStateChanges = observabilityScope.trap({
-            try self.computePackageStateChanges(
-                root: root,
-                resolvedDependencies: updateResults,
-                updateBranches: updateBranches,
-                observabilityScope: observabilityScope
-            )
-        }) else {
-            return []
-        }
-
-        // First remove the checkouts that are no longer required.
-        for (packageRef, state) in packageStateChanges {
-            observabilityScope.makeChildScope(description: "removing unneeded checkouts", metadata: packageRef.diagnosticsMetadata).trap {
-                switch state {
-                case .added, .updated, .unchanged: 
-                    break
-                case .removed:
-                    try self.remove(package: packageRef)
-                }
-            }
-        }
-
-        // Update or clone new packages.
-        for (packageRef, state) in packageStateChanges {
-            observabilityScope.makeChildScope(description: "updating or cloning new packages", metadata: packageRef.diagnosticsMetadata).trap {
-                switch state {
-                case .added(let state):
-                    _ = try self.updateDependency(package: packageRef, requirement: state.requirement, productFilter: state.products, observabilityScope: observabilityScope)
-                case .updated(let state):
-                    _ = try self.updateDependency(package: packageRef, requirement: state.requirement, productFilter: state.products, observabilityScope: observabilityScope)
-                case .removed, .unchanged:
-                    break
-                }
-            }
-        }
-
-        // Inform the delegate if nothing was updated.
-        if packageStateChanges.filter({ $0.1 == .unchanged }).count == packageStateChanges.count {
-            delegate?.dependenciesUpToDate()
-        }
-
-        return packageStateChanges
-    }
-
-    private func updateDependency(
-        package: PackageReference,
-        requirement: PackageStateChange.Requirement,
-        productFilter: ProductFilter,
-        observabilityScope: ObservabilityScope
-    ) throws -> AbsolutePath {
-        switch requirement {
-        case .version(let version):
-            // FIXME: this should not block
-            let container = try temp_await {
-                packageContainerProvider.getContainer(
-                    for: package,
-                    updateStrategy: .never,
-                    observabilityScope: observabilityScope,
-                    on: .sharedConcurrent,
-                    completion: $0
-                )
-            }
-
-            if let container = container as? SourceControlPackageContainer {
-                // FIXME: We need to get the revision here, and we don't have a
-                // way to get it back out of the resolver which is very
-                // annoying. Maybe we should make an SPI on the provider for this?
-                guard let tag = container.getTag(for: version) else {
-                    throw InternalError("unable to get tag for \(package) \(version); available versions \(try container.versionsDescending())")
-                }
-                let revision = try container.getRevision(forTag: tag)
-                try container.checkIntegrity(version: version, revision: revision)
-                return try self.checkoutRepository(package: package, at: .version(version, revision: revision), observabilityScope: observabilityScope)
-            } else if let _ = container as? RegistryPackageContainer {
-                return try self.downloadRegistryArchive(package: package, at: version, observabilityScope: observabilityScope)
-            } else if let customContainer = container as? CustomPackageContainer {
-                let path = try customContainer.retrieve(at: version, observabilityScope: observabilityScope)
-                let dependency = try ManagedDependency(packageRef: package, state: .custom(version: version, path: path), subpath: RelativePath(validating: ""))
-                self.state.dependencies.add(dependency)
-                try self.state.save()
-                return path
-            } else {
-                throw InternalError("invalid container for \(package.identity) of type \(package.kind)")
-            }
-
-        case .revision(let revision, .none):
-            return try self.checkoutRepository(package: package, at: .revision(revision), observabilityScope: observabilityScope)
-
-        case .revision(let revision, .some(let branch)):
-            return try self.checkoutRepository(package: package, at: .branch(name: branch, revision: revision), observabilityScope: observabilityScope)
-
-        case .unversioned:
-            let dependency = try ManagedDependency.fileSystem(packageRef: package)
-            // this is silly since we just created it above, but no good way to force cast it and extract the path
-            guard case .fileSystem(let path) = dependency.state else {
-                throw InternalError("invalid package type: \(package.kind)")
-            }
-
-            self.state.dependencies.add(dependency)
-            try self.state.save()
-            return path
-        }
-    }
-
-    public enum ResolutionPrecomputationResult: Equatable {
-        case required(reason: WorkspaceResolveReason)
-        case notRequired
-
-        public var isRequired: Bool {
-            switch self {
-            case .required: return true
-            case .notRequired: return false
-            }
-        }
-    }
-
-    /// Computes if dependency resolution is required based on input constraints and pins.
-    ///
-    /// - Returns: Returns a result defining whether dependency resolution is required and the reason for it.
-    // @testable internal
-    public func precomputeResolution(
-        root: PackageGraphRoot,
-        dependencyManifests: DependencyManifests,
-        pinsStore: PinsStore,
-        constraints: [PackageContainerConstraint],
-        observabilityScope: ObservabilityScope
-    ) throws -> ResolutionPrecomputationResult {
-        let computedConstraints =
-            try root.constraints() +
-            // Include constraints from the manifests in the graph root.
-            root.manifests.values.flatMap{ try $0.dependencyConstraints(productFilter: .everything) } +
-            dependencyManifests.dependencyConstraints +
-            constraints
-
-        let precomputationProvider = ResolverPrecomputationProvider(root: root, dependencyManifests: dependencyManifests)
-        let resolver = PubGrubDependencyResolver(provider: precomputationProvider, pins: pinsStore.pins, observabilityScope: observabilityScope)
-        let result = resolver.solve(constraints: computedConstraints)
-
-        guard !observabilityScope.errorsReported else {
-            return .required(reason: .errorsPreviouslyReported)
-        }
-
-        switch result {
-        case .success:
-            return .notRequired
-        case .failure(ResolverPrecomputationError.missingPackage(let package)):
-            return .required(reason: .newPackages(packages: [package]))
-        case .failure(ResolverPrecomputationError.differentRequirement(let package, let state, let requirement)):
-            return .required(reason: .packageRequirementChange(
-                package: package,
-                state: state,
-                requirement: requirement
-            ))
-        case .failure(let error):
-            return .required(reason: .other("\(error.interpolationDescription)"))
-        }
-    }
-
-    /// Validates that each checked out managed dependency has an entry in pinsStore.
-    private func loadAndUpdatePinsStore(
-        dependencyManifests: DependencyManifests,
-        rootManifestsMinimumToolsVersion: ToolsVersion,
-        observabilityScope: ObservabilityScope
-    ) -> PinsStore?  {
-        guard let pinsStore = observabilityScope.trap({ try self.pinsStore.load() }) else {
-            return nil
-        }
-
-        guard let requiredDependencies = observabilityScope.trap({ try dependencyManifests.requiredPackages.filter({ $0.kind.isPinnable }) }) else {
-            return nil
-        }
-        for dependency in self.state.dependencies.filter({ $0.packageRef.kind.isPinnable }) {
-            // a required dependency that is already loaded (managed) should be represented in the pins store.
-            // also comparing location as it may have changed at this point
-            if requiredDependencies.contains(where: { $0.equalsIncludingLocation(dependency.packageRef) }) {
-                let pin = pinsStore.pins[dependency.packageRef.identity]
-                // if pin not found, or location is different (it may have changed at this point) pin it
-                if !(pin?.packageRef.equalsIncludingLocation(dependency.packageRef) ?? false) {
-                    pinsStore.pin(dependency)
-                }
-            } else if let pin = pinsStore.pins[dependency.packageRef.identity]  {
-                // otherwise, it should *not* be in the pins store.
-                pinsStore.remove(pin)
-            }
-        }
-
-        return pinsStore
-    }
-
-    /// This enum represents state of an external package.
-    public enum PackageStateChange: Equatable, CustomStringConvertible {
-
-        /// The requirement imposed by the the state.
-        public enum Requirement: Equatable, CustomStringConvertible {
-            /// A version requirement.
-            case version(Version)
-
-            /// A revision requirement.
-            case revision(Revision, branch: String?)
-
-            case unversioned
-
-            public var description: String {
-                switch self {
-                case .version(let version):
-                    return "requirement(\(version))"
-                case .revision(let revision, let branch):
-                    return "requirement(\(revision) \(branch ?? ""))"
-                case .unversioned:
-                    return "requirement(unversioned)"
-                }
-            }
-
-            public var prettyPrinted: String {
-                switch self {
-                case .version(let version):
-                    return "\(version)"
-                case .revision(let revision, let branch):
-                    return "\(revision) \(branch ?? "")"
-                case .unversioned:
-                    return "unversioned"
-                }
-            }
-        }
-        public struct State: Equatable {
-            public let requirement: Requirement
-            public let products: ProductFilter
-            public init(requirement: Requirement, products: ProductFilter) {
-                self.requirement = requirement
-                self.products = products
-            }
-        }
-
-        /// The package is added.
-        case added(State)
-
-        /// The package is removed.
-        case removed
-
-        /// The package is unchanged.
-        case unchanged
-
-        /// The package is updated.
-        case updated(State)
-
-        public var description: String {
-            switch self {
-            case .added(let requirement):
-                return "added(\(requirement))"
-            case .removed:
-                return "removed"
-            case .unchanged:
-                return "unchanged"
-            case .updated(let requirement):
-                return "updated(\(requirement))"
-            }
-        }
-
-        public var isAddedOrUpdated: Bool {
-            switch self {
-            case .added, .updated:
-                return true
-            case .unchanged, .removed:
-                return false
-            }
-        }
-    }
-
-    /// Computes states of the packages based on last stored state.
-    fileprivate func computePackageStateChanges(
-        root: PackageGraphRoot,
-        resolvedDependencies: [(PackageReference, BoundVersion, ProductFilter)],
-        updateBranches: Bool,
-        observabilityScope: ObservabilityScope
-    ) throws -> [(PackageReference, PackageStateChange)] {
-        // Load pins store and managed dependencies.
-        let pinsStore = try self.pinsStore.load()
-        var packageStateChanges: [PackageIdentity: (PackageReference, PackageStateChange)] = [:]
-
-        // Set the states from resolved dependencies results.
-        for (packageRef, binding, products) in resolvedDependencies {
-            // Get the existing managed dependency for this package ref, if any.
-
-            // first find by identity only since edit location may be different by design
-            var currentDependency = self.state.dependencies[packageRef.identity]
-            // Check if this is an edited dependency.
-            if case .edited(let basedOn, _) = currentDependency?.state, let originalReference = basedOn?.packageRef {
-                packageStateChanges[originalReference.identity] = (originalReference, .unchanged)
-            } else {
-                // if not edited, also compare by location since it may have changed
-                currentDependency = self.state.dependencies[comparingLocation: packageRef]
-            }
-
-            switch binding {
-            case .excluded:
-                throw InternalError("Unexpected excluded binding")
-
-            case .unversioned:
-                // Ignore the root packages.
-                if root.packages.keys.contains(packageRef.identity) {
-                    continue
-                }
-
-                if let currentDependency {
-                    switch currentDependency.state {
-                    case .fileSystem, .edited:
-                        packageStateChanges[packageRef.identity] = (packageRef, .unchanged)
-                    case .sourceControlCheckout:
-                        let newState = PackageStateChange.State(requirement: .unversioned, products: products)
-                        packageStateChanges[packageRef.identity] = (packageRef, .updated(newState))
-                    case .registryDownload:
-                        throw InternalError("Unexpected unversioned binding for downloaded dependency")
-                    case .custom:
-                        throw InternalError("Unexpected unversioned binding for custom dependency")
-                    }
-                } else {
-                    let newState = PackageStateChange.State(requirement: .unversioned, products: products)
-                    packageStateChanges[packageRef.identity] = (packageRef, .added(newState))
-                }
-
-            case .revision(let identifier, let branch):
-                // Get the latest revision from the container.
-                // TODO: replace with async/await when available
-                guard let container = (try temp_await {
-                    packageContainerProvider.getContainer(
-                        for: packageRef,
-                        updateStrategy: .never,
-                        observabilityScope: observabilityScope,
-                        on: .sharedConcurrent,
-                        completion: $0
-                    )
-                }) as? SourceControlPackageContainer else {
-                    throw InternalError("invalid container for \(packageRef) expected a SourceControlPackageContainer")
-                }
-                var revision = try container.getRevision(forIdentifier: identifier)
-                let branch = branch ?? (identifier == revision.identifier ? nil : identifier)
-
-                // If we have a branch and we shouldn't be updating the
-                // branches, use the revision from pin instead (if present).
-                if branch != nil, !updateBranches {
-                    if case .branch(branch, let pinRevision) = pinsStore.pins.values.first(where: { $0.packageRef == packageRef })?.state {
-                        revision = Revision(identifier: pinRevision)
-                    }
-                }
-
-                // First check if we have this dependency.
-                if let currentDependency {
-                    // If current state and new state are equal, we don't need
-                    // to do anything.
-                    let newState: CheckoutState
-                    if let branch {
-                        newState = .branch(name: branch, revision: revision)
-                    } else {
-                        newState = .revision(revision)
-                    }
-                    if case .sourceControlCheckout(let checkoutState) = currentDependency.state, checkoutState == newState {
-                        packageStateChanges[packageRef.identity] = (packageRef, .unchanged)
-                    } else {
-                        // Otherwise, we need to update this dependency to this revision.
-                        let newState = PackageStateChange.State(requirement: .revision(revision, branch: branch), products: products)
-                        packageStateChanges[packageRef.identity] = (packageRef, .updated(newState))
-                    }
-                } else {
-                    let newState = PackageStateChange.State(requirement: .revision(revision, branch: branch), products: products)
-                    packageStateChanges[packageRef.identity] = (packageRef, .added(newState))
-                }
-
-            case .version(let version):
-                let stateChange: PackageStateChange
-                switch currentDependency?.state {
-                case .sourceControlCheckout(.version(version, _)), .registryDownload(version), .custom(version, _):
-                    stateChange = .unchanged
-                case .edited, .fileSystem, .sourceControlCheckout, .registryDownload, .custom:
-                    stateChange = .updated(.init(requirement: .version(version), products: products))
-                case nil:
-                    stateChange = .added(.init(requirement: .version(version), products: products))
-                }
-                packageStateChanges[packageRef.identity] = (packageRef, stateChange)
-            }
-        }
-        // Set the state of any old package that might have been removed.
-        for packageRef in self.state.dependencies.lazy.map({ $0.packageRef }) where packageStateChanges[packageRef.identity] == nil {
-            packageStateChanges[packageRef.identity] = (packageRef, .removed)
-        }
-
-        return Array(packageStateChanges.values)
-    }
-
-    /// Creates resolver for the workspace.
-    fileprivate func createResolver(pins: PinsStore.Pins, observabilityScope: ObservabilityScope) throws -> PubGrubDependencyResolver {
-        var delegate: DependencyResolverDelegate
-        let observabilityDelegate = ObservabilityDependencyResolverDelegate(observabilityScope: observabilityScope)
-        if let workspaceDelegate = self.delegate {
-            delegate = MultiplexResolverDelegate([
-                observabilityDelegate,
-                WorkspaceDependencyResolverDelegate(workspaceDelegate),
-            ])
-        } else {
-            delegate = observabilityDelegate
-        }
-
-        return PubGrubDependencyResolver(
-            provider: packageContainerProvider,
-            pins: pins,
-            skipDependenciesUpdates: self.configuration.skipDependenciesUpdates,
-            prefetchBasedOnResolvedFile: self.configuration.prefetchBasedOnResolvedFile,
-            observabilityScope: observabilityScope,
-            delegate: delegate
-        )
-    }
-
-    /// Runs the dependency resolver based on constraints provided and returns the results.
-    fileprivate func resolveDependencies(
-        resolver: PubGrubDependencyResolver,
-        constraints: [PackageContainerConstraint],
-        observabilityScope: ObservabilityScope
-    ) -> [(package: PackageReference, binding: BoundVersion, products: ProductFilter)] {
-
-        os_signpost(.begin, name: SignpostName.pubgrub)
-        let result = resolver.solve(constraints: constraints)
-        os_signpost(.end, name: SignpostName.pubgrub)
-
-        // Take an action based on the result.
-        switch result {
-        case .success(let bindings):
-            return bindings
-        case .failure(let error):
-            observabilityScope.emit(error)
-            return []
-        }
-    }
-}
-
-
-
-/// A result which can be loaded.
-///
-/// It is useful for objects that holds a state on disk and needs to be
-/// loaded frequently.
-public final class LoadableResult<Value> {
-    /// The constructor closure for the value.
-    private let construct: () throws -> Value
-
-    /// Create a loadable result.
-    public init(_ construct: @escaping () throws -> Value) {
-        self.construct = construct
-    }
-
-    /// Load and return the result.
-    public func loadResult() -> Result<Value, Error> {
-        return Result(catching: {
-            try self.construct()
-        })
-    }
-
-    /// Load and return the value.
-    public func load() throws -> Value {
-        return try loadResult().get()
-    }
-}
-
-// MARK: - Package container provider
-
-extension Workspace: PackageContainerProvider {
-    public func getContainer(
-        for package: PackageReference,
-        updateStrategy: ContainerUpdateStrategy,
-        observabilityScope: ObservabilityScope,
-        on queue: DispatchQueue,
-        completion: @escaping (Result<PackageContainer, Swift.Error>) -> Void
+    public func acceptIdentityChange(
+       package: PackageIdentity,
+       version: Version,
+       signingEntity: SigningEntity,
+       origin: SigningEntity.Origin,
+       observabilityScope: ObservabilityScope,
+       callbackQueue: DispatchQueue,
+       completion: @escaping (Result<Void, Error>) -> Void
     ) {
-        do {
-            switch package.kind {
-            // If the container is local, just create and return a local package container.
-            case .root, .fileSystem:
-                let container = try FileSystemPackageContainer(
-                    package: package,
-                    identityResolver: self.identityResolver,
-                    dependencyMapper: self.dependencyMapper,
-                    manifestLoader: self.manifestLoader,
-                    currentToolsVersion: self.currentToolsVersion,
-                    fileSystem: self.fileSystem,
-                    observabilityScope: observabilityScope
-                )
-                queue.async {
-                    completion(.success(container))
-                }
-            // Resolve the container using the repository manager.
-            case .localSourceControl, .remoteSourceControl:
-                let repositorySpecifier = try package.makeRepositorySpecifier()
-                self.repositoryManager.lookup(
-                    package: package.identity,
-                    repository: repositorySpecifier,
-                    updateStrategy: updateStrategy.repositoryUpdateStrategy,
-                    observabilityScope: observabilityScope,
-                    delegateQueue: queue,
-                    callbackQueue: queue
-                ) { result in
-                    dispatchPrecondition(condition: .onQueue(queue))
-                    // Create the container wrapper.
-                    let result = result.tryMap { handle -> PackageContainer in
-                        // Open the repository.
-                        //
-                        // FIXME: Do we care about holding this open for the lifetime of the container.
-                        let repository = try handle.open()
-                        return try SourceControlPackageContainer(
-                            package: package,
-                            identityResolver: self.identityResolver,
-                            dependencyMapper: self.dependencyMapper,
-                            repositorySpecifier: repositorySpecifier,
-                            repository: repository,
-                            manifestLoader: self.manifestLoader,
-                            currentToolsVersion: self.currentToolsVersion,
-                            fingerprintStorage: self.fingerprints,
-                            fingerprintCheckingMode: FingerprintCheckingMode.map(self.configuration.fingerprintCheckingMode),
-                            observabilityScope: observabilityScope
-                        )
-                    }
-                    completion(result)
-                }
-            // Resolve the container using the registry
-            case .registry:
-                let container = RegistryPackageContainer(
-                    package: package,
-                    identityResolver: self.identityResolver,
-                    dependencyMapper: self.dependencyMapper,
-                    registryClient: self.registryClient,
-                    manifestLoader: self.manifestLoader,
-                    currentToolsVersion: self.currentToolsVersion,
-                    observabilityScope: observabilityScope
-                )
-                queue.async {
-                    completion(.success(container))
-                }
-            }
-        } catch {
-            queue.async {
-                completion(.failure(error))
-            }
-        }
+        self.registryClient.changeSigningEntityFromVersion(
+           package: package,
+           version: version,
+           signingEntity: signingEntity,
+           origin: origin,
+           observabilityScope: observabilityScope,
+           callbackQueue: callbackQueue,
+           completion: completion
+        )
     }
+}
 
+extension Workspace {
     /// Removes the clone and checkout of the provided specifier.
     ///
     /// - Parameters:
@@ -2955,354 +1430,6 @@ extension Workspace: PackageContainerProvider {
     }
 }
 
-extension FingerprintCheckingMode {
-    static func map(_ checkingMode: WorkspaceConfiguration.CheckingMode) -> FingerprintCheckingMode {
-        switch checkingMode {
-        case .strict:
-            return .strict
-        case .warn:
-            return .warn
-        }
-    }
-}
-
-extension SigningEntityCheckingMode {
-    static func map(_ checkingMode: WorkspaceConfiguration.CheckingMode) -> SigningEntityCheckingMode {
-        switch checkingMode {
-        case .strict:
-            return .strict
-        case .warn:
-            return .warn
-        }
-    }
-}
-
-// MARK: - Source control repository management
-
-// FIXME: this mixes quite a bit of workspace logic with repository specific one
-// need to better separate the concerns
-extension Workspace {
-    /// Create a local clone of the given `repository` checked out to `checkoutState`.
-    ///
-    /// If an existing clone is present, the repository will be reset to the
-    /// requested revision, if necessary.
-    ///
-    /// - Parameters:
-    ///   - package: The package to clone.
-    ///   - checkoutState: The state to check out.
-    /// - Returns: The path of the local repository.
-    /// - Throws: If the operation could not be satisfied.
-    func checkoutRepository(
-        package: PackageReference,
-        at checkoutState: CheckoutState,
-        observabilityScope: ObservabilityScope
-    ) throws -> AbsolutePath {
-        let repository = try package.makeRepositorySpecifier()
-
-        // first fetch the repository
-        let checkoutPath = try self.fetchRepository(
-            package: package,
-            at: checkoutState.revision,
-            observabilityScope: observabilityScope
-        )
-
-        // Check out the given revision.
-        let workingCopy = try self.repositoryManager.openWorkingCopy(at: checkoutPath)
-
-        // Inform the delegate that we're about to start.
-        delegate?.willCheckOut(package: package.identity, repository: repository.location.description, revision: checkoutState.description, at: checkoutPath)
-        let start = DispatchTime.now()
-
-        // Do mutable-immutable dance because checkout operation modifies the disk state.
-        try fileSystem.chmod(.userWritable, path: checkoutPath, options: [.recursive, .onlyFiles])
-        try workingCopy.checkout(revision: checkoutState.revision)
-        try? fileSystem.chmod(.userUnWritable, path: checkoutPath, options: [.recursive, .onlyFiles])
-
-        // Record the new state.
-        observabilityScope.emit(debug: "adding '\(package.identity)' (\(package.locationString)) to managed dependencies", metadata: package.diagnosticsMetadata)
-        self.state.dependencies.add(
-            try .sourceControlCheckout(
-                packageRef: package,
-                state: checkoutState,
-                subpath: checkoutPath.relative(to: self.location.repositoriesCheckoutsDirectory)
-            )
-        )
-        try self.state.save()
-
-        // Inform the delegate that we're done.
-        let duration = start.distance(to: .now())
-        delegate?.didCheckOut(package: package.identity, repository: repository.location.description, revision: checkoutState.description, at: checkoutPath, duration: duration)
-        observabilityScope.emit(debug: "`\(repository.location.description)` checked out at \(checkoutState.debugDescription)")
-
-        return checkoutPath
-    }
-
-    func checkoutRepository(
-        package: PackageReference,
-        at pinState: PinsStore.PinState,
-        observabilityScope: ObservabilityScope
-    ) throws -> AbsolutePath {
-        switch pinState {
-        case .version(let version, revision: let revision) where revision != nil:
-            return try self.checkoutRepository(
-                package: package,
-                at: .version(version, revision: .init(identifier: revision!)), // nil checked above
-                observabilityScope: observabilityScope
-            )
-        case .branch(let branch, revision: let revision):
-            return try self.checkoutRepository(
-                package: package,
-                at: .branch(name: branch, revision: .init(identifier: revision)),
-                observabilityScope: observabilityScope
-            )
-        case .revision(let revision):
-            return try self.checkoutRepository(
-                package: package,
-                at: .revision(.init(identifier: revision)),
-                observabilityScope: observabilityScope
-            )
-        default:
-            throw InternalError("invalid pin state: \(pinState)")
-        }
-    }
-
-    /// Fetch a given `package` and create a local checkout for it.
-    ///
-    /// This will first clone the repository into the canonical repositories
-    /// location, if necessary, and then check it out from there.
-    ///
-    /// - Returns: The path of the local repository.
-    /// - Throws: If the operation could not be satisfied.
-    private func fetchRepository(
-        package: PackageReference,
-        at revision: Revision,
-        observabilityScope: ObservabilityScope
-    ) throws -> AbsolutePath {
-        let repository = try package.makeRepositorySpecifier()
-
-        // If we already have it, fetch to update the repo from its remote.
-        // also compare the location as it may have changed
-        if let dependency = self.state.dependencies[comparingLocation: package] {
-            let checkoutPath = self.location.repositoriesCheckoutSubdirectory(for: dependency)
-
-            // Make sure the directory is not missing (we will have to clone again if not).
-            // This can become invalid if the build directory is moved.
-            fetch: if self.fileSystem.isDirectory(checkoutPath) {
-                // Fetch the checkout in case there are updates available.
-                let workingCopy = try self.repositoryManager.openWorkingCopy(at: checkoutPath)
-
-                // Ensure that the alternative object store is still valid.
-                guard try self.repositoryManager.isValidWorkingCopy(workingCopy, for: repository) else {
-                    observabilityScope.emit(debug: "working copy at '\(checkoutPath)' does not align with expected local path of '\(repository)'")
-                    break fetch
-                }
-
-                // only update if necessary
-                if !workingCopy.exists(revision: revision) {
-                    // The fetch operation may update contents of the checkout, 
-                    // so we need to do mutable-immutable dance.
-                    try self.fileSystem.chmod(.userWritable, path: checkoutPath, options: [.recursive, .onlyFiles])
-                    try workingCopy.fetch()
-                    try? self.fileSystem.chmod(.userUnWritable, path: checkoutPath, options: [.recursive, .onlyFiles])
-                }
-
-                return checkoutPath
-            }
-        }
-
-        // If not, we need to get the repository from the checkouts.
-        // FIXME: this should not block
-        let handle = try temp_await {
-            self.repositoryManager.lookup(
-                package: package.identity,
-                repository: repository,
-                updateStrategy: .never,
-                observabilityScope: observabilityScope,
-                delegateQueue: .sharedConcurrent,
-                callbackQueue: .sharedConcurrent,
-                completion: $0
-            )
-        }
-
-        // Clone the repository into the checkouts.
-        let checkoutPath = self.location.repositoriesCheckoutsDirectory.appending(component: repository.basename)
-
-        // Remove any existing content at that path.
-        try self.fileSystem.chmod(.userWritable, path: checkoutPath, options: [.recursive, .onlyFiles])
-        try self.fileSystem.removeFileTree(checkoutPath)
-
-        // Inform the delegate that we're about to start.
-        self.delegate?.willCreateWorkingCopy(package: package.identity, repository: handle.repository.location.description, at: checkoutPath)
-        let start = DispatchTime.now()
-        
-        // Create the working copy.
-        _ = try handle.createWorkingCopy(at: checkoutPath, editable: false)
-
-        // Inform the delegate that we're done.
-        let duration = start.distance(to: .now())
-        self.delegate?.didCreateWorkingCopy(package: package.identity, repository: handle.repository.location.description, at: checkoutPath, duration: duration)
-
-        return checkoutPath
-    }
-
-    /// Removes the clone and checkout of the provided specifier.
-    fileprivate func removeRepository(dependency: ManagedDependency) throws {
-        guard case .sourceControlCheckout = dependency.state else {
-            throw InternalError("cannot remove repository for \(dependency) with state \(dependency.state)")
-        }
-
-        // Remove the checkout.
-        let dependencyPath = self.location.repositoriesCheckoutSubdirectory(for: dependency)
-        let workingCopy = try self.repositoryManager.openWorkingCopy(at: dependencyPath)
-        guard !workingCopy.hasUncommittedChanges() else {
-            throw WorkspaceDiagnostics.UncommitedChanges(repositoryPath: dependencyPath)
-        }
-
-        try self.fileSystem.chmod(.userWritable, path: dependencyPath, options: [.recursive, .onlyFiles])
-        try self.fileSystem.removeFileTree(dependencyPath)
-
-        // Remove the clone.
-        try self.repositoryManager.remove(repository: dependency.packageRef.makeRepositorySpecifier())
-    }
-}
-
-// MARK: - Registry Source archive management
-
- extension Workspace {
-     func downloadRegistryArchive(
-        package: PackageReference,
-        at version: Version,
-        observabilityScope: ObservabilityScope
-     ) throws -> AbsolutePath {
-         // FIXME: this should not block
-         let downloadPath = try temp_await {
-             self.registryDownloadsManager.lookup(
-                package: package.identity,
-                version: version,
-                observabilityScope: observabilityScope,
-                delegateQueue: .sharedConcurrent,
-                callbackQueue: .sharedConcurrent,
-                completion: $0
-             )
-         }
-
-         // Record the new state.
-         observabilityScope.emit(debug: "adding '\(package.identity)' (\(package.locationString)) to managed dependencies", metadata: package.diagnosticsMetadata)
-         self.state.dependencies.add(
-            try .registryDownload(
-                packageRef: package,
-                version: version,
-                subpath: downloadPath.relative(to: self.location.registryDownloadDirectory)
-            )
-         )
-         try self.state.save()
-
-         return downloadPath
-     }
-
-     func downloadRegistryArchive(
-        package: PackageReference,
-        at pinState: PinsStore.PinState,
-        observabilityScope: ObservabilityScope
-     ) throws -> AbsolutePath {
-         switch pinState {
-         case .version(let version, _):
-             return try self.downloadRegistryArchive(
-                package: package,
-                at: version,
-                observabilityScope: observabilityScope
-             )
-         default:
-             throw InternalError("invalid pin state: \(pinState)")
-         }
-     }
-
-     func removeRegistryArchive(for dependency: ManagedDependency) throws {
-         guard case .registryDownload = dependency.state else {
-             throw InternalError("cannot remove source archive for \(dependency) with state \(dependency.state)")
-         }
-
-         let downloadPath = self.location.registryDownloadSubdirectory(for: dependency)
-         try self.fileSystem.removeFileTree(downloadPath)
-
-         // remove the local copy
-         try registryDownloadsManager.remove(package: dependency.packageRef.identity)
-     }
-
-     public func acceptIdentityChange(
-        package: PackageIdentity,
-        version: Version,
-        signingEntity: SigningEntity,
-        origin: SigningEntity.Origin,
-        observabilityScope: ObservabilityScope,
-        callbackQueue: DispatchQueue,
-        completion: @escaping (Result<Void, Error>) -> Void
-     ) {
-         self.registryClient.changeSigningEntityFromVersion(
-            package: package,
-            version: version,
-            signingEntity: signingEntity,
-            origin: origin,
-            observabilityScope: observabilityScope,
-            callbackQueue: callbackQueue,
-            completion: completion
-         )
-     }
- }
-
-// MARK: - Signatures
-
-extension Workspace {
-    private func validateSignatures(
-        packageGraph: PackageGraph,
-        expectedSigningEntities: [PackageIdentity: RegistryReleaseMetadata.SigningEntity]
-    ) throws {
-        try expectedSigningEntities.forEach { identity, expectedSigningEntity in
-            if let package = packageGraph.packages.first(where: { $0.identity == identity }) {
-                guard let actualSigningEntity = package.registryMetadata?.signature?.signedBy else {
-                    throw SigningError.unsigned(package: identity, expected: expectedSigningEntity)
-                }
-                if actualSigningEntity != expectedSigningEntity {
-                    throw SigningError.mismatchedSigningEntity(
-                        package: identity,
-                        expected: expectedSigningEntity,
-                        actual: actualSigningEntity
-                    )
-                }
-            } else {
-                guard let mirror = self.mirrors.mirror(for: identity.description) else {
-                    throw SigningError.expectedIdentityNotFound(package: identity)
-                }
-                let mirroredIdentity = PackageIdentity.plain(mirror)
-                guard mirroredIdentity.isRegistry else {
-                    throw SigningError.expectedSignedMirroredToSourceControl(package: identity, expected: expectedSigningEntity)
-                }
-                guard let package = packageGraph.packages.first(where: { $0.identity == mirroredIdentity }) else {
-                    // Unsure if this case is reachable in practice.
-                    throw SigningError.expectedIdentityNotFound(package: identity)
-                }
-                guard let actualSigningEntity = package.registryMetadata?.signature?.signedBy else {
-                    throw SigningError.unsigned(package: identity, expected: expectedSigningEntity)
-                }
-                if actualSigningEntity != expectedSigningEntity {
-                    throw SigningError.mismatchedSigningEntity(
-                        package: identity,
-                        expected: expectedSigningEntity,
-                        actual: actualSigningEntity
-                    )
-                }
-            }
-        }
-    }
-    
-    public enum SigningError: Swift.Error {
-        case expectedIdentityNotFound(package: PackageIdentity)
-        case expectedSignedMirroredToSourceControl(package: PackageIdentity, expected: RegistryReleaseMetadata.SigningEntity)
-        case mismatchedSigningEntity(package: PackageIdentity, expected: RegistryReleaseMetadata.SigningEntity, actual: RegistryReleaseMetadata.SigningEntity)
-        case unsigned(package: PackageIdentity, expected: RegistryReleaseMetadata.SigningEntity)
-    }
-}
-
 // MARK: - Utility extensions
 
 fileprivate extension Workspace.ManagedArtifact {
@@ -3316,28 +1443,6 @@ fileprivate extension Workspace.ManagedArtifact {
     }
 }
 
-// FIXME: the manifest loading logic should be changed to use identity instead of location once identity is unique
-// at that time we should remove this
-//@available(*, deprecated)
-fileprivate extension PackageDependency {
-    var locationString: String {
-        switch self {
-        case .fileSystem(let settings):
-            return settings.path.pathString
-        case .sourceControl(let settings):
-            switch settings.location {
-            case .local(let path):
-                return path.pathString
-            case .remote(let url):
-                return url.absoluteString
-            }
-        case .registry:
-            // FIXME: placeholder
-            return self.identity.description
-        }
-    }
-}
-
 internal extension PackageReference {
     func makeRepositorySpecifier() throws -> RepositorySpecifier {
         switch self.kind {
@@ -3347,17 +1452,6 @@ internal extension PackageReference {
             return .init(url: url)
         default:
             throw StringError("invalid dependency kind \(self.kind)")
-        }
-    }
-}
-
-fileprivate extension PackageReference.Kind {
-    var isPinnable: Bool {
-        switch self {
-        case .remoteSourceControl, .localSourceControl, .registry:
-            return true
-        default:
-            return false
         }
     }
 }
@@ -3436,65 +1530,6 @@ extension Workspace {
     }
 }
 
-extension PinsStore.PinState {
-    fileprivate func equals(_ checkoutState: CheckoutState) -> Bool {
-        switch (self, checkoutState) {
-        case (.version(let lversion, let lrevision), .version(let rversion, let rrevision)):
-            return lversion == rversion && lrevision == rrevision.identifier
-        case (.branch(let lbranch, let lrevision), .branch(let rbranch, let rrevision)):
-            return lbranch == rbranch && lrevision == rrevision.identifier
-        case (.revision(let lrevision), .revision(let rrevision)):
-            return lrevision == rrevision.identifier
-        default:
-            return false
-        }
-    }
-
-    fileprivate func equals(_ version: Version) -> Bool {
-        switch self {
-        case .version(let version, _):
-            return version == version
-        default:
-            return false
-        }
-    }
-}
-
-extension CheckoutState {
-    fileprivate var revision: Revision {
-        get {
-            switch self {
-            case .revision(let revision):
-                return revision
-            case .version(_, let revision):
-                return revision
-            case .branch(_, let revision):
-                return revision
-            }
-        }
-    }
-
-    var isBranchOrRevisionBased: Bool {
-        switch self {
-        case .revision, .branch:
-            return true
-        case .version:
-            return false
-        }
-    }
-
-    fileprivate var requirement: PackageRequirement {
-        switch self {
-        case .revision(let revision):
-            return .revision(revision.identifier)
-        case .version(let version, _):
-            return .versionSet(.exact(version))
-        case .branch(let branch, _):
-            return .revision(branch)
-        }
-    }
-}
-
 extension Workspace.Location {
     /// Returns the path to the dependency's repository checkout directory.
     func repositoriesCheckoutSubdirectory(for dependency: Workspace.ManagedDependency) -> AbsolutePath {
@@ -3512,35 +1547,6 @@ extension Workspace.Location {
     }
 }
 
-extension FileSystem {
-    // helper to decide if an archive directory would benefit from stripping first level
-    fileprivate func shouldStripFirstLevel(archiveDirectory: AbsolutePath, acceptableExtensions: [String]? = nil) throws -> Bool {
-        let subdirectories = try self.getDirectoryContents(archiveDirectory)
-            .map{ archiveDirectory.appending(component: $0) }
-            .filter { self.isDirectory($0) }
-
-        // single top-level directory required
-        guard subdirectories.count == 1, let rootDirectory = subdirectories.first else {
-            return false
-        }
-
-        // no acceptable extensions defined, so the single top-level directory is a good candidate
-        guard let acceptableExtensions else {
-            return true
-        }
-
-        // the single top-level directory is already one of the acceptable extensions, so no need to strip
-        if rootDirectory.extension.map({ acceptableExtensions.contains($0) }) ?? false {
-            return false
-        }
-
-        // see if there is "grand-child" directory with one of the acceptable extensions
-        return try self.getDirectoryContents(rootDirectory)
-            .map{ rootDirectory.appending(component: $0) }
-            .first{ $0.extension.map { acceptableExtensions.contains($0) } ?? false } != nil
-    }
-}
-
 extension Workspace.Location {
     func validatingSharedLocations(
         fileSystem: FileSystem,
@@ -3552,7 +1558,9 @@ extension Workspace.Location {
             keyPath: \.sharedConfigurationDirectory,
             fileSystem: fileSystem,
             getOrCreateHandler: {
-                try fileSystem.getOrCreateSwiftPMConfigurationDirectory(warningHandler: self.emitDeprecatedConfigurationWarning ? warningHandler : { _ in })
+                try fileSystem.getOrCreateSwiftPMConfigurationDirectory(
+                    warningHandler: self.emitDeprecatedConfigurationWarning ? warningHandler : { _ in }
+                )
             },
             warningHandler: warningHandler
         )
@@ -3607,317 +1615,6 @@ extension Workspace.Location {
     }
 }
 
-extension Workspace {
-    // the goal of this code is to help align dependency identities across source control and registry origins
-    // the issue this solves is that dependencies will have different identities across the origins
-    // for example, source control based dependency on http://github.com/apple/swift-nio would have an identifier of "swift-nio"
-    // while in the registry, the same package will [likely] have an identifier of "apple.swift-nio"
-    // since there is not generally fire sure way to translate one system to the other (urls can vary widely, so the best we would be able to do is guess)
-    // what this code does is query the registry of it "knows" what the registry identity of URL is, and then use the registry identity instead of the URL bases one
-    // the code also supports a "full swizzle" mode in which it _replaces_ the source control dependency with a registry one which encourages the transition
-    // from source control based dependencies to registry based ones
-
-    // TODO
-    // 1. handle mixed situation when some versions on the registry but some on source control. we need a second lookup to make sure the version exists
-    // 2. handle registry returning multiple identifiers, how do we choose the right one?
-    fileprivate struct RegistryAwareManifestLoader: ManifestLoaderProtocol {
-        private let underlying: ManifestLoaderProtocol
-        private let registryClient: RegistryClient
-        private let transformationMode: TransformationMode
-
-        private let cacheTTL = DispatchTimeInterval.seconds(300) // 5m
-        private let identityLookupCache = ThreadSafeKeyValueStore<SourceControlURL, (result: Result<PackageIdentity?, Error>, expirationTime: DispatchTime)>()
-
-        init(underlying: ManifestLoaderProtocol, registryClient: RegistryClient, transformationMode: TransformationMode) {
-            self.underlying = underlying
-            self.registryClient = registryClient
-            self.transformationMode = transformationMode
-        }
-
-        func load(
-            manifestPath: AbsolutePath,
-            manifestToolsVersion: ToolsVersion,
-            packageIdentity: PackageIdentity,
-            packageKind: PackageReference.Kind,
-            packageLocation: String,
-            packageVersion: (version: Version?, revision: String?)?,
-            identityResolver: IdentityResolver,
-            dependencyMapper: DependencyMapper,
-            fileSystem: FileSystem,
-            observabilityScope: ObservabilityScope,
-            delegateQueue: DispatchQueue,
-            callbackQueue: DispatchQueue,
-            completion: @escaping (Result<Manifest, Error>) -> Void
-        ) {
-            self.underlying.load(
-                manifestPath: manifestPath,
-                manifestToolsVersion: manifestToolsVersion,
-                packageIdentity: packageIdentity,
-                packageKind: packageKind,
-                packageLocation: packageLocation,
-                packageVersion: packageVersion,
-                identityResolver: identityResolver,
-                dependencyMapper: dependencyMapper,
-                fileSystem: fileSystem,
-                observabilityScope: observabilityScope,
-                delegateQueue: delegateQueue,
-                callbackQueue: callbackQueue
-            ) { result in
-                switch result {
-                case .failure(let error):
-                    completion(.failure(error))
-                case .success(let manifest):
-                    self.transformSourceControlDependenciesToRegistry(
-                        manifest: manifest,
-                        transformationMode: transformationMode,
-                        observabilityScope: observabilityScope,
-                        callbackQueue: callbackQueue,
-                        completion: completion
-                    )
-                }
-            }
-        }
-
-        func resetCache(observabilityScope: ObservabilityScope) {
-            self.underlying.resetCache(observabilityScope: observabilityScope)
-        }
-
-        func purgeCache(observabilityScope: ObservabilityScope) {
-            self.underlying.purgeCache(observabilityScope: observabilityScope)
-        }
-
-        private func transformSourceControlDependenciesToRegistry(
-            manifest: Manifest,
-            transformationMode: TransformationMode,
-            observabilityScope: ObservabilityScope,
-            callbackQueue: DispatchQueue,
-            completion: @escaping (Result<Manifest, Error>) -> Void
-        ) {
-            let sync = DispatchGroup()
-            let transformations = ThreadSafeKeyValueStore<PackageDependency, PackageIdentity>()
-            for dependency in manifest.dependencies {
-                if case .sourceControl(let settings) = dependency, case .remote(let url) = settings.location  {
-                    sync.enter()
-                    self.mapRegistryIdentity(url: url, observabilityScope: observabilityScope, callbackQueue: callbackQueue) { result in
-                        defer { sync.leave() }
-                        switch result {
-                        case .failure(let error):
-                            // do not raise error, only report it as warning
-                            observabilityScope.emit(
-                                warning: "failed querying registry identity for '\(url)'",
-                                underlyingError: error
-                            )
-                        case .success(.some(let identity)):
-                            transformations[dependency] = identity
-                        case .success(.none):
-                            // no identity found
-                            break
-                        }
-                    }
-                }
-            }
-
-            // update the manifest with the transformed dependencies
-            sync.notify(queue: callbackQueue) {
-                do {
-                    let updatedManifest = try self.transformManifest(
-                        manifest: manifest,
-                        transformations: transformations.get(),
-                        transformationMode: transformationMode,
-                        observabilityScope: observabilityScope
-                    )
-                    completion(.success(updatedManifest))
-                }
-                catch {
-                    return completion(.failure(error))
-                }
-            }
-        }
-
-        private func transformManifest(
-            manifest: Manifest,
-            transformations: [PackageDependency: PackageIdentity],
-            transformationMode: TransformationMode,
-            observabilityScope: ObservabilityScope
-        ) throws -> Manifest {
-            var targetDependencyPackageNameTransformations = [String: String]()
-
-            var modifiedDependencies = [PackageDependency]()
-            for dependency in manifest.dependencies {
-                var modifiedDependency = dependency
-                if let registryIdentity = transformations[dependency] {
-                    guard case .sourceControl(let settings) = dependency, case .remote = settings.location else {
-                        // an implementation mistake
-                        throw InternalError("unexpected non-source-control dependency: \(dependency)")
-                    }
-                    switch transformationMode {
-                    case .identity:
-                        // we replace the *identity* of the dependency in order to align the identities
-                        // and de-dupe across source control and registry origins
-                        observabilityScope.emit(info: "adjusting '\(dependency.locationString)' identity to registry identity of '\(registryIdentity)'.")
-                        modifiedDependency = .sourceControl(
-                            identity: registryIdentity,
-                            nameForTargetDependencyResolutionOnly: settings.nameForTargetDependencyResolutionOnly,
-                            location: settings.location,
-                            requirement: settings.requirement,
-                            productFilter: settings.productFilter
-                        )
-                    case .swizzle:
-                        // we replace the *entire* source control dependency with a registry one
-                        // this helps de-dupe across source control and registry dependencies
-                        // and also encourages use of registry over source control
-                        switch settings.requirement {
-                        case .exact, .range:
-                            let requirement = try settings.requirement.asRegistryRequirement()
-                            observabilityScope.emit(info: "swizzling '\(dependency.locationString)' with registry dependency '\(registryIdentity)'.")
-                            targetDependencyPackageNameTransformations[dependency.nameForTargetDependencyResolutionOnly] = registryIdentity.description
-                            modifiedDependency = .registry(
-                                identity: registryIdentity,
-                                requirement: requirement,
-                                productFilter: settings.productFilter
-                            )
-                        case .branch, .revision:
-                            // branch and revision dependencies are not supported by the registry
-                            // in such case, the best we can do is to replace the *identity* of the
-                            // source control dependency in order to align the identities
-                            // and de-dupe across source control and registry origins
-                            observabilityScope.emit(info: "adjusting '\(dependency.locationString)' identity to registry identity of '\(registryIdentity)'.")
-                            modifiedDependency = .sourceControl(
-                                identity: registryIdentity,
-                                nameForTargetDependencyResolutionOnly: settings.nameForTargetDependencyResolutionOnly,
-                                location: settings.location,
-                                requirement: settings.requirement,
-                                productFilter: settings.productFilter
-                            )
-                        }
-                    }
-                }
-                modifiedDependencies.append(modifiedDependency)
-            }
-
-            var modifiedTargets = manifest.targets
-            if !transformations.isEmpty {
-                modifiedTargets = []
-                for target in manifest.targets {
-                    var modifiedDependencies = [TargetDescription.Dependency]()
-                    for dependency in target.dependencies {
-                        var modifiedDependency = dependency
-                        if case .product(let name, let packageName, let moduleAliases, let condition) = dependency, let packageName = packageName {
-                            // makes sure we use the updated package name for target based dependencies
-                            if let modifiedPackageName = targetDependencyPackageNameTransformations[packageName] {
-                                modifiedDependency = .product(name: name, package: modifiedPackageName, moduleAliases: moduleAliases, condition: condition)
-                            }
-                        }
-                        modifiedDependencies.append(modifiedDependency)
-                    }
-
-                    modifiedTargets.append(
-                        try TargetDescription(
-                            name: target.name,
-                            dependencies: modifiedDependencies,
-                            path: target.path,
-                            url: target.url,
-                            exclude: target.exclude,
-                            sources: target.sources,
-                            resources: target.resources,
-                            publicHeadersPath: target.publicHeadersPath,
-                            type: target.type,
-                            packageAccess: target.packageAccess,
-                            pkgConfig: target.pkgConfig,
-                            providers: target.providers,
-                            pluginCapability: target.pluginCapability,
-                            settings: target.settings,
-                            checksum: target.checksum,
-                            pluginUsages: target.pluginUsages
-                        )
-                    )
-                }
-            }
-
-            let modifiedManifest = Manifest(
-                displayName: manifest.displayName,
-                path: manifest.path,
-                packageKind: manifest.packageKind,
-                packageLocation: manifest.packageLocation,
-                defaultLocalization: manifest.defaultLocalization,
-                platforms: manifest.platforms,
-                version: manifest.version,
-                revision: manifest.revision,
-                toolsVersion: manifest.toolsVersion,
-                pkgConfig: manifest.pkgConfig,
-                providers: manifest.providers,
-                cLanguageStandard: manifest.cLanguageStandard,
-                cxxLanguageStandard: manifest.cxxLanguageStandard,
-                swiftLanguageVersions: manifest.swiftLanguageVersions,
-                dependencies: modifiedDependencies,
-                products: manifest.products,
-                targets: modifiedTargets
-            )
-
-            return modifiedManifest
-        }
-
-        private func mapRegistryIdentity(
-            url: SourceControlURL,
-            observabilityScope: ObservabilityScope,
-            callbackQueue: DispatchQueue,
-            completion: @escaping (Result<PackageIdentity?, Error>) -> Void
-        ) {
-            if let cached = self.identityLookupCache[url], cached.expirationTime > .now() {
-                switch cached.result {
-                case .success(let identity):
-                    return completion(.success(identity))
-                case .failure:
-                    // server error, do not try again
-                    return completion(.success(.none))
-                }
-            }
-
-            self.registryClient.lookupIdentities(scmURL: url, observabilityScope: observabilityScope, callbackQueue: callbackQueue) { result in
-                switch result {
-                case .failure(let error):
-                    self.identityLookupCache[url] = (result: .failure(error), expirationTime: .now() + self.cacheTTL)
-                    completion(.failure(error))
-                case .success(let identities):
-                    // FIXME: returns first result... need to consider how to address multiple ones
-                    let identity = identities.sorted().first
-                    self.identityLookupCache[url] = (result: .success(identity), expirationTime: .now() + self.cacheTTL)
-                    completion(.success(identity))
-                }
-            }
-        }
-
-        enum TransformationMode {
-            case identity
-            case swizzle
-
-            init?(_ seed: WorkspaceConfiguration.SourceControlToRegistryDependencyTransformation) {
-                switch seed {
-                case .identity:
-                    self = .identity
-                case .swizzle:
-                    self = .swizzle
-                case .disabled:
-                    return nil
-                }
-            }
-        }
-    }
-}
-
-fileprivate extension PackageDependency.SourceControl.Requirement {
-    func asRegistryRequirement() throws -> PackageDependency.Registry.Requirement {
-        switch self {
-        case .range(let versions):
-            return .range(versions)
-        case .exact(let version):
-            return .exact(version)
-        case .branch, .revision:
-            throw InternalError("invalid source control to registry requirement tranformation")
-        }
-    }
-}
-
 fileprivate func warnToStderr(_ message: String) {
     TSCBasic.stderrStream.write("warning: \(message)\n")
     TSCBasic.stderrStream.flush()
@@ -3936,18 +1633,5 @@ extension ContainerUpdateStrategy {
         case .ifNeeded(let revision):
             return .ifNeeded(revision: .init(identifier: revision))
         }
-    }
-}
-
-extension Workspace.ManagedDependencies {
-    func hasEditedDependencies() -> Bool {
-        self.contains(where: {
-            switch $0.state {
-            case .edited(_, _):
-                return true
-            default:
-                return false
-            }
-        })
     }
 }
